+++ src/appleseedmaya/exporters/lightexporter.cpp
@@ -34,6 +34,7 @@
 
 // Maya headers.
 #include <maya/MFnDagNode.h>
+#include "appleseedmaya/mayaheaderscleanup.h"
 
 // appleseed.foundation headers.
 #include "foundation/math/scalar.h"
@@ -151,7 +152,7 @@ void LightExporter::createEntities(
 
         MAngle penumbraAngle(5.0f, MAngle::kDegrees);
         AttributeUtils::get(node(), "penumbraAngle", penumbraAngle);
-        const float outerAngle = coneAngle.asDegrees() + 2.0 * penumbraAngle.asDegrees();
+        const double outerAngle = coneAngle.asDegrees() + 2.0 * penumbraAngle.asDegrees();
         lightParams.insert("outer_angle", outerAngle);
     }
     else
