+++ src/appleseedmaya/exporters/mandelbrotexporter.cpp
@@ -36,6 +36,7 @@
 
 // Maya headers.
 #include <maya/MFnDependencyNode.h>
+#include "appleseedmaya/mayaheaderscleanup.h"
 
 // appleseed.renderer headers.
 #include "renderer/api/utility.h"
@@ -127,7 +128,7 @@ void MandelbrotExporter::exportParameterValue(
         std::vector<MandelbrotColorsEntry> mandelbrotColors;
         mandelbrotColors.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
@@ -175,7 +176,7 @@ void MandelbrotExporter::exportParameterValue(
         std::vector<MandelbrotValuesEntry> mandelbrotValues;
         mandelbrotValues.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
