+++ src/appleseedmaya/exporters/remapcolorexporter.cpp
@@ -36,6 +36,7 @@
 
 // Maya headers.
 #include <maya/MFnDependencyNode.h>
+#include "appleseedmaya/mayaheaderscleanup.h"
 
 // appleseed.renderer headers.
 #include "renderer/api/utility.h"
@@ -108,7 +109,7 @@ void RemapColorExporter::exportParameterValue(
         std::vector<RemapColorEntry> remapRed;
         remapRed.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
@@ -154,7 +155,7 @@ void RemapColorExporter::exportParameterValue(
         std::vector<RemapColorEntry> remapGreen;
         remapGreen.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
@@ -200,7 +201,7 @@ void RemapColorExporter::exportParameterValue(
         std::vector<RemapColorEntry> remapBlue;
         remapBlue.reserve(plug.numElements());
 
-        for(size_t i = 0, e = plug.numElements(); i < e; ++i)
+        for(unsigned int i = 0, e = plug.numElements(); i < e; ++i)
         {
             MPlug entry = plug.elementByPhysicalIndex(i);
             MPlug position = entry.child(0);
